+++ ta4j-core/src/main/java/org/ta4j/core/indicators/aroon/AroonUpIndicator.java
@@ -21,75 +21,79 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators;
+package org.ta4j.core.indicators.aroon;
 
 import static org.ta4j.core.num.NaN.NaN;
 
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.indicators.helpers.LowPriceIndicator;
-import org.ta4j.core.indicators.helpers.LowestValueIndicator;
+import org.ta4j.core.indicators.CachedIndicator;
+import org.ta4j.core.indicators.helpers.HighPriceIndicator;
+import org.ta4j.core.indicators.helpers.HighestValueIndicator;
 import org.ta4j.core.num.Num;
 
 /**
- * Aroon down indicator.
+ * Aroon up indicator.
  *
  * @see <a href=
  *      "http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:aroon">chart_school:technical_indicators:aroon</a>
  */
-public class AroonDownIndicator extends CachedIndicator<Num> {
+public class AroonUpIndicator extends CachedIndicator<Num> {
 
     private final int barCount;
-    private final LowestValueIndicator lowestLowPriceIndicator;
-    private final Indicator<Num> lowPriceIndicator;
+    private final HighestValueIndicator highestHighPriceIndicator;
+    private final Indicator<Num> highPriceIndicator;
     private final Num hundred;
+    private final Num barCountNum;
 
     /**
      * Constructor.
      *
-     * @param lowPriceIndicator the indicator for the low price (default
-     *                          {@link LowPriceIndicator})
-     * @param barCount          the time frame
+     * @param highPriceIndicator the indicator for the high price (default
+     *                           {@link HighPriceIndicator})
+     * @param barCount           the time frame
      */
-    public AroonDownIndicator(Indicator<Num> lowPriceIndicator, int barCount) {
-        super(lowPriceIndicator);
+    public AroonUpIndicator(Indicator<Num> highPriceIndicator, int barCount) {
+        super(highPriceIndicator);
         this.barCount = barCount;
-        this.lowPriceIndicator = lowPriceIndicator;
+        this.highPriceIndicator = highPriceIndicator;
         this.hundred = numOf(100);
+        this.barCountNum = numOf(barCount);
         // + 1 needed for last possible iteration in loop
-        this.lowestLowPriceIndicator = new LowestValueIndicator(lowPriceIndicator, barCount + 1);
+        this.highestHighPriceIndicator = new HighestValueIndicator(highPriceIndicator, barCount + 1);
     }
 
     /**
-     * Default Constructor that is using the low price
+     * Default Constructor that is using the high price
      *
      * @param series   the bar series
      * @param barCount the time frame
      */
-    public AroonDownIndicator(BarSeries series, int barCount) {
-        this(new LowPriceIndicator(series), barCount);
+    public AroonUpIndicator(BarSeries series, int barCount) {
+        this(new HighPriceIndicator(series), barCount);
     }
 
     @Override
     protected Num calculate(int index) {
-        if (getBarSeries().getBar(index).getLowPrice().isNaN())
+        if (getBarSeries().getBar(index).getHighPrice().isNaN())
             return NaN;
 
-        // Getting the number of bars since the lowest close price
+        // Getting the number of bars since the highest close price
         int endIndex = Math.max(0, index - barCount);
         int nbBars = 0;
         for (int i = index; i > endIndex; i--) {
-            if (lowPriceIndicator.getValue(i).isEqual(lowestLowPriceIndicator.getValue(index))) {
+            if (highPriceIndicator.getValue(i).isEqual(highestHighPriceIndicator.getValue(index))) {
                 break;
             }
             nbBars++;
         }
 
-        return numOf(barCount - nbBars).dividedBy(numOf(barCount)).multipliedBy(hundred);
+        return numOf(barCount - nbBars).dividedBy(barCountNum).multipliedBy(hundred);
     }
 
     @Override
     public String toString() {
         return getClass().getSimpleName() + " barCount: " + barCount;
     }
+
 }